+++ app/__init__.py
@@ -77,6 +77,8 @@ def main():
     app.config["SECRET_KEY"] = SECRET_KEY
     app.config["SQLALCHEMY_DATABASE_URI"] = f"sqlite:///{app.root_path}/{DB_NAME}"
     app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+    app.config["SESSION_COOKIE_SAMESITE"] = 'Lax'
+    app.config["REMEMBER_COOKIE_SAMESITE"] = 'Lax'
     app.logger.removeHandler(default_handler)
     migrate = Migrate(app, db, render_as_batch=True)